+++ node/src/types/kernel/api/block/foldBlock/payload.d.ts
@@ -16,21 +16,14 @@
  */
 
 //#region content
-/* directories */
-export * as appendBlock from "./appendBlock";
-export * as deleteBlock from "./deleteBlock";
-export * as foldBlock from "./foldBlock";
-export * as getBlockBreadcrumb from "./getBlockBreadcrumb";
-export * as getBlockDOM from "./getBlockDOM";
-export * as getBlockInfo from "./getBlockInfo";
-export * as getBlockKramdown from "./getBlockKramdown";
-export * as getChildBlocks from "./getChildBlocks";
-export * as getDocInfo from "./getDocInfo";
-export * as insertBlock from "./insertBlock";
-export * as moveBlock from "./moveBlock";
-export * as prependBlock from "./prependBlock";
-export * as transferBlockRef from "./transferBlockRef";
-export * as updateBlock from "./updateBlock";
+/**
+ * Fold a block
+ */
+export interface IPayload {
+    /**
+     * ID of the block to be folded
+     */
+    id: string;
+}
 
-/* files */
 //#endregion content